Migrated to Microk8s
This commit migrates this project to using Microk8s by modifying
the hostpath volume mount paths so that they may be mounted using
Microk8s. This required that the paths be hardcoded and hence all
the paths in this commit need to be changed if these kuberenetes
manifests are used for deployment from another folder.

Migration to Microk8s was done even though minikube allowd
mounting hostpaths and chaning their location within the cluster,
because Minio failed to run with paths whoes location was
changed.